Add some more options to latencytop
Add the ability to set a specific runtime if we are looking at a live system.
Also allow us to filter for only a specific task name, which is helpful when
debugging a big multithreaded app.  Also keep track of total time we were
tracing and use that as the percentage so we have a true look at how long we
were asleep for the total time, rather than as a percentage of the sleep time.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>