Keep better track of how long we are asleep
Especially in collapse mode we can have a bunch of tasks all sleeping at
overlapping times which makes for a sleeptime larger than the time we traced.
So instead keep track of when we go to sleep and when we wake up and then add
together those unique times for a better view of total sleep time wrt the total
sample time.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>