Better output formatting and waketime latency tracking
Get a better idea of how long it takes a task to be woken up from when it was
told to wake up by tracking min/avg/max per process when doing wake tracking.
Also adjust the output a little since we now have a whole bunch of new stuff to
show.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>