Add the ability to trace a specific command
It's helpful to run and profile just one specific command, so add the -r option
to do just that.  We still have to filter the ftrace output a bit for some
reason, will have to figure out why the set_ftrace_pid filtering doesn't do all
the work we want to do but thats a job for future-josef.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>